+++ Assets/Statistics/Scripts/DataRecord.cs
@@ -7,10 +7,10 @@ namespace Statistics
     public class DataRecord
     {
         public float time;
-        public int money;
+        public float money;
         public int units;
         
-        public DataRecord(float time, int money, int units)
+        public DataRecord(float time, float money, int units)
         {
             this.time = time;
             this.money = money;
